feat: add test for cardList
