+++ react-components/.eslintrc
@@ -1,57 +1,65 @@
 {
-    "extends": [
-      "eslint:recommended",
-      "plugin:@typescript-eslint/eslint-recommended",
-      "plugin:@typescript-eslint/recommended",
-      "plugin:react/recommended",
-      "plugin:react-hooks/recommended"
-    ],
-    "rules": {
-      "eqeqeq": 1,
-      "react/prop-types": 0,
-      "react/react-in-jsx-scope": 0,
-      "react-hooks/rules-of-hooks": "off",
-      "react-hooks/exhaustive-deps": "warn",
-      "no-console": ["warn", { "allow": ["error"] }],
-      "no-debugger": "warn",
-      "require-await": "error",
-      "no-undef": "error",
-      "no-param-reassign": "error",
-      "prefer-const": "error",
-      "no-duplicate-imports": "error",
-      "@typescript-eslint/no-unused-vars": [
-        "warn",
-        {
-          "args": "after-used",
-          "ignoreRestSiblings": true
-        }
-      ],
-      "no-use-before-define": "off",
-      "@typescript-eslint/no-explicit-any": "off",
-      "@typescript-eslint/no-empty-function": "off",
-      "@typescript-eslint/explicit-module-boundary-types": "off"
-    },
-    "parser": "@typescript-eslint/parser",
-    "parserOptions": {
-      "ecmaFeatures": {
-        "experimentalObjectRestSpread": true,
-        "jsx": true
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:react/recommended",
+    "plugin:react-hooks/recommended"
+  ],
+  "rules": {
+    "eqeqeq": 1,
+    "react/prop-types": 0,
+    "react/react-in-jsx-scope": 0,
+    "react-hooks/rules-of-hooks": "off",
+    "react-hooks/exhaustive-deps": "warn",
+    "no-console": ["warn", { "allow": ["error"] }],
+    "no-debugger": "warn",
+    "require-await": "error",
+    "no-undef": "error",
+    "no-param-reassign": "error",
+    "prefer-const": "error",
+    "no-duplicate-imports": "error",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      {
+        "args": "after-used",
+        "ignoreRestSiblings": true
       }
-    },
-    "env": {
-      "browser": true,
-      "es6": true,
-      "node": true
-    },
-    "globals": {
-      "history": false,
-      "document": true,
-      "JSX": true
-    },
-    "settings": {
-      "react": {
-        "pragma": "React",
-        "version": "detect"
+    ],
+    "no-use-before-define": "off",
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-empty-function": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off"
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaFeatures": {
+      "experimentalObjectRestSpread": true,
+      "jsx": true
+    }
+  },
+  "env": {
+    "browser": true,
+    "es6": true,
+    "node": true
+  },
+  "globals": {
+    "history": false,
+    "document": true,
+    "JSX": true
+  },
+  "settings": {
+    "react": {
+      "pragma": "React",
+      "version": "detect"
+    }
+  },
+  "overrides": [
+    {
+      "files": ["**/*.test.ts", "**/*.test.tsx"],
+      "env": {
+        "jest": true
       }
     }
-  }
+  ]
+}
