+++ react-components/src/react-components.test/cardList.test.tsx
@@ -1,0 +1,18 @@
+import { render, screen } from '@testing-library/react'
+import React from 'react'
+
+import Card from '../components/Card'
+
+import { allProducts } from '../constants/allProducts'
+
+describe('CardList', () => {
+  it('render list', () => {
+    class CardList extends React.Component {
+      render() {
+        return allProducts.map((item) => <Card key={item.id} {...item} />)
+      }
+    }
+    render(<CardList />)
+    expect(screen.getAllByTestId('card').length).toEqual(20)
+  })
+})
