+++ react-components/vite.config.ts
@@ -11,5 +11,8 @@ export default defineConfig({
     globals: true,
     environment: 'jsdom',
     setupFiles: './src/tests/setup.ts',
+    coverage: {
+      all: true,
+    },
   },
 })
